requirements.txt
flask==2.2.5
flask-sqlalchemy==3.0.5
flask-login==0.6.2
flask-jwt-extended==4.5.2
flask-caching==2.0.2
flask-profiler==1.8.1
flask-admin==1.6.1
python-dotenv==1.0.0
requests==2.31.0
mysqlclient==2.1.1
redis==4.5.4
celery==5.3.1
prometheus-flask-exporter==0.23.2
structlog==25.2.0
sentry-sdk==2.24.1
transformers==4.50.1
torch
numpy
scikit-learn
spacy>=3.2.3,<3.3.0
ja-ginza>=5.1.0,<6.0.0
passlib==1.7.4
bcrypt==3.2.0
python-jose==3.3.0
aiohttp==3.8.1
asyncio==3.4.3
prometheus-client==0.16.0
memory-profiler==0.61.0
sqlalchemy-utils==0.41.1 